Moved export service out of exam package as it shrunk in size.  Added comment for clarification of exportExam.
