+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/service/ArchiveExamExportService.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.processor.service.export;
+package org.opentestsystem.rdw.reporting.processor.service;
 
 import org.opentestsystem.rdw.reporting.common.report.UserPermissions;
 import org.opentestsystem.rdw.reporting.processor.model.ExamExportQuery;
@@ -34,6 +34,8 @@ public class ArchiveExamExportService implements ExamExportService {
 
     @Override
     public String exportExam(final UserPermissions permissions, final ExamExportQuery query) {
+        // Only FileSchemePrefix needs to get replaced for local MySql.
+        // Aurora's SELECT INTO OUTFILE S3 feature shares the same s3:// as archive's ResourceLoader.
         String uriLocation = archiveRoot.replace(FileSchemePrefix, "/") + exportBasePath;
         String fileNamePrefix = String.format(FileNameFormat, dateFormat.format(new Date()));
