+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/service/ExamExportService.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.reporting.processor.service.export;
+package org.opentestsystem.rdw.reporting.processor.service;
 
 import org.opentestsystem.rdw.reporting.common.report.UserPermissions;
 import org.opentestsystem.rdw.reporting.processor.model.ExamExportQuery;
