+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/ExportController.java
@@ -3,7 +3,7 @@ package org.opentestsystem.rdw.reporting.processor.web;
 import com.google.common.collect.ImmutableMap;
 import org.opentestsystem.rdw.reporting.common.report.UserPermissions;
 import org.opentestsystem.rdw.reporting.processor.model.ExamExportQuery;
-import org.opentestsystem.rdw.reporting.processor.service.export.ExamExportService;
+import org.opentestsystem.rdw.reporting.processor.service.ExamExportService;
 import org.opentestsystem.rdw.security.Permission;
 import org.opentestsystem.rdw.security.PermissionScope;
 import org.springframework.beans.factory.annotation.Autowired;
